Reworked item table
